Change string formatting verb to %v from %w
This commit changes string formatting verb to %v from %w since %w verb
can only be used once. This was also done because %w is used to make the
make the returned error from fmt.Errorf be available to errors.Is and
errors.As, but these are not even used in the go-bitbucket package.