Update branch restrictions response struct and decode func
This commit reverts back the branch restrictions to just house the
client and adds a branch restrictions response object that also contains
a slice of branch restrictions returned by the api response. The decode
function for branch restrictions is updated to act similarly to the diff
stat decoder in that it decodes the top level of the json response
into the branch restrictions response struct.